Many changes
Fixed a how pilot assigns triggers because it was broken.

Add a method to create trigger zones inside of swerve so we can do things only when in a certain part of the field.

Also restructed auton to one file, we shouldn't need to modify Robot as much for auton things.
PathPlanner event triggers are setup at the top, and then you use them in RobotSTates to define what they do.

Modified Field to give us a few other helpful methods.

And updated Phoenix-6 library

drive locks to 45s when driving and steering, but not if steering only.