Refactored RobotConfig so it's not needed
Rio.java lets us setup which controller we are on as an enum, and is easier to work with than RobotConfig.

Config is now built into the Robot.java file like it is for most other systems and makes adding in new subsystems easier.

Still need to test this on an actual robot before committing to main