Changes from 2/16/2025
Build a format for changes on the coming vision file:
with inspiration from 2024 code, 2025 vision commands will have similar functions but more related to reef alignment

Any changes to AlignToVisionTarget are discontinued and will be worked out and only used as a rudimentary guide to building configs and PIDcontrollers in vision.java

Future:
Maybe plans to include alignment to cage?
or alignment to source/human player coral station (hpcs)?

Co-Authored-By: marcjuguilon <133045955+marcjuguilon@users.noreply.github.com>